requirements.txt
mylogging>=4.0.0
mydatapreprocessing>=2.0.9
EelForkExcludeFiles
black
build
GitPython
nbmake
pandas
phmdoctest
plotly
pyinstaller
pytest
pytest-cov
setuptools
sphinx
tabulate
twine
typeguard
virtualenv
wheel